Small changes to bring this branch closer to master.

+++ src/core/modules/Skylight.cpp
@@ -139,8 +139,8 @@ void Skylight::setParamsv(const float * _sunPos, float _turbidity)
 	// Set the two main variables
 	//thetas = M_PI_2 - std::asin((float)sunPos[2]);
 	// GZ why not directly acos?
-	thetas = std::acosf(sunPos[2]);
-	T = _turbidity;
+	thetas = std::acos(static_cast<double>(sunPos[2]));
+	T = static_cast<double>(_turbidity);
 
 	// Precomputation of the distribution coefficients and zenith luminances/color
 	computeZenithLuminance();
@@ -149,7 +149,7 @@ void Skylight::setParamsv(const float * _sunPos, float _turbidity)
 	computeColorDistributionCoefs();
 
 	// Precompute everything possible to increase the get_CIE_value() function speed
-	float cos_thetas = sunPos[2];
+	double cos_thetas = static_cast<double>(sunPos[2]);
 	term_x = zenithColorX   / ((1. + Ax * std::exp(Bx)) * (1. + Cx * std::exp(Dx*thetas) + Ex * cos_thetas * cos_thetas));
 	term_y = zenithColorY   / ((1. + Ay * std::exp(By)) * (1. + Cy * std::exp(Dy*thetas) + Ey * cos_thetas * cos_thetas));
 	term_Y = 0.0; // zenithLuminance /((1.f + AY * std::exp(BY)) * (1.f + CY * std::exp(DY*thetas) + EY * cos_thetas * cos_thetas));
